requirements.txt
awsebcli==3.14.1
backports.ssl-match-hostname==3.5.0.1
blessed==1.14.2
boto3==1.7.39
botocore==1.10.37
cached-property==1.4.2
cement==2.8.2
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
Django==2.0.1
django-datatables-view==1.16.0
django-debug-toolbar==1.9.1
django-discover-runner==1.0
django-role-permissions==2.2.0
django-storages==1.6.6
django-widget-tweaks==1.4.2
docker==3.3.0
docker-compose==1.21.2
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
idna==2.6
jmespath==0.9.3
jsonschema==2.6.0
mysqlclient==1.4.2.post1
pathspec==0.5.5
phonenumbers==8.9.2
Pillow==6.0.0
python-dateutil==2.7.3
python-magic==0.4.15
pytz==2018.3
PyYAML==3.12
requests==2.18.4
s3transfer==0.1.13
semantic-version==2.5.0
six==1.11.0
sqlparse==0.2.4
tabulate==0.7.5
termcolor==1.1.0
texttable==0.9.1
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.48.0
